Refactor our Resource.getListHTML to utility class
