+++ jetty-core/jetty-util/src/main/java/org/eclipse/jetty/util/resource/Resource.java
@@ -32,11 +32,9 @@ import java.nio.file.ProviderNotFoundException;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.StandardOpenOption;
 import java.nio.file.attribute.FileTime;
-import java.text.DateFormat;
 import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Collection;
-import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Objects;
@@ -47,8 +45,6 @@ import org.eclipse.jetty.util.FileID;
 import org.eclipse.jetty.util.IO;
 import org.eclipse.jetty.util.Index;
 import org.eclipse.jetty.util.Loader;
-import org.eclipse.jetty.util.MultiMap;
-import org.eclipse.jetty.util.StringUtil;
 import org.eclipse.jetty.util.URIUtil;
 import org.eclipse.jetty.util.UrlEncoded;
 import org.slf4j.Logger;
@@ -731,236 +727,6 @@ public abstract class Resource implements ResourceFactory
     }
 
     /**
-     * Get the resource list as a HTML directory listing.
-     *
-     * @param base The base URL
-     * @param parent True if the parent directory should be included
-     * @param query query params
-     * @return String of HTML
-     */
-    public String getListHTML(String base, boolean parent, String query)  // TODO: move to helper class
-    {
-        // This method doesn't check aliases, so it is OK to canonicalize here.
-        base = URIUtil.normalizePath(base);
-        if (base == null || !isDirectory())
-            return null;
-
-        List<String> rawListing = list();
-        if (rawListing == null)
-        {
-            return null;
-        }
-
-        boolean sortOrderAscending = true;
-        String sortColumn = "N"; // name (or "M" for Last Modified, or "S" for Size)
-
-        // check for query
-        if (query != null)
-        {
-            MultiMap<String> params = new MultiMap<>();
-            UrlEncoded.decodeUtf8To(query, 0, query.length(), params);
-
-            String paramO = params.getString("O");
-            String paramC = params.getString("C");
-            if (StringUtil.isNotBlank(paramO))
-            {
-                if (paramO.equals("A"))
-                {
-                    sortOrderAscending = true;
-                }
-                else if (paramO.equals("D"))
-                {
-                    sortOrderAscending = false;
-                }
-            }
-            if (StringUtil.isNotBlank(paramC))
-            {
-                if (paramC.equals("N") || paramC.equals("M") || paramC.equals("S"))
-                {
-                    sortColumn = paramC;
-                }
-            }
-        }
-
-        // Gather up entries
-        List<Resource> items = new ArrayList<>();
-        for (String l : rawListing)
-        {
-            // TODO review the re-encoding here.  Perhaps better to get list from path
-            Resource item = resolve(URIUtil.encodePath(l));
-            items.add(item);
-        }
-
-        // Perform sort
-        if (sortColumn.equals("M"))
-        {
-            items.sort(ResourceCollators.byLastModified(sortOrderAscending));
-        }
-        else if (sortColumn.equals("S"))
-        {
-            items.sort(ResourceCollators.bySize(sortOrderAscending));
-        }
-        else
-        {
-            items.sort(ResourceCollators.byName(sortOrderAscending));
-        }
-
-        String decodedBase = URIUtil.decodePath(base);
-        String title = "Directory: " + deTag(decodedBase);
-
-        StringBuilder buf = new StringBuilder(4096);
-
-        // Doctype Declaration (HTML5)
-        buf.append("<!DOCTYPE html>\n");
-        buf.append("<html lang=\"en\">\n");
-
-        // HTML Header
-        buf.append("<head>\n");
-        buf.append("<meta charset=\"utf-8\">\n");
-        buf.append("<link href=\"jetty-dir.css\" rel=\"stylesheet\" />\n");
-        buf.append("<title>");
-        buf.append(title);
-        buf.append("</title>\n");
-        buf.append("</head>\n");
-
-        // HTML Body
-        buf.append("<body>\n");
-        buf.append("<h1 class=\"title\">").append(title).append("</h1>\n");
-
-        // HTML Table
-        final String ARROW_DOWN = "&nbsp; &#8681;";
-        final String ARROW_UP = "&nbsp; &#8679;";
-
-        buf.append("<table class=\"listing\">\n");
-        buf.append("<thead>\n");
-
-        String arrow = "";
-        String order = "A";
-        if (sortColumn.equals("N"))
-        {
-            if (sortOrderAscending)
-            {
-                order = "D";
-                arrow = ARROW_UP;
-            }
-            else
-            {
-                order = "A";
-                arrow = ARROW_DOWN;
-            }
-        }
-
-        buf.append("<tr><th class=\"name\"><a href=\"?C=N&O=").append(order).append("\">");
-        buf.append("Name").append(arrow);
-        buf.append("</a></th>");
-
-        arrow = "";
-        order = "A";
-        if (sortColumn.equals("M"))
-        {
-            if (sortOrderAscending)
-            {
-                order = "D";
-                arrow = ARROW_UP;
-            }
-            else
-            {
-                order = "A";
-                arrow = ARROW_DOWN;
-            }
-        }
-
-        buf.append("<th class=\"lastmodified\"><a href=\"?C=M&O=").append(order).append("\">");
-        buf.append("Last Modified").append(arrow);
-        buf.append("</a></th>");
-
-        arrow = "";
-        order = "A";
-        if (sortColumn.equals("S"))
-        {
-            if (sortOrderAscending)
-            {
-                order = "D";
-                arrow = ARROW_UP;
-            }
-            else
-            {
-                order = "A";
-                arrow = ARROW_DOWN;
-            }
-        }
-        buf.append("<th class=\"size\"><a href=\"?C=S&O=").append(order).append("\">");
-        buf.append("Size").append(arrow);
-        buf.append("</a></th></tr>\n");
-        buf.append("</thead>\n");
-
-        buf.append("<tbody>\n");
-
-        String encodedBase = hrefEncodeURI(base);
-
-        if (parent)
-        {
-            // Name
-            buf.append("<tr><td class=\"name\"><a href=\"");
-            buf.append(URIUtil.addPaths(encodedBase, "../"));
-            buf.append("\">Parent Directory</a></td>");
-            // Last Modified
-            buf.append("<td class=\"lastmodified\">-</td>");
-            // Size
-            buf.append("<td>-</td>");
-            buf.append("</tr>\n");
-        }
-
-        DateFormat dfmt = DateFormat.getDateTimeInstance(DateFormat.MEDIUM,
-            DateFormat.MEDIUM);
-        for (Resource item : items)
-        {
-            String name = item.getFileName();
-            if (StringUtil.isBlank(name))
-            {
-                continue; // skip
-            }
-
-            if (item.isDirectory() && !name.endsWith("/"))
-            {
-                name += URIUtil.SLASH;
-            }
-
-            // Name
-            buf.append("<tr><td class=\"name\"><a href=\"");
-            String path = URIUtil.addEncodedPaths(encodedBase, URIUtil.encodePath(name));
-            buf.append(path);
-            buf.append("\">");
-            buf.append(deTag(name));
-            buf.append("&nbsp;");
-            buf.append("</a></td>");
-
-            // Last Modified
-            buf.append("<td class=\"lastmodified\">");
-            long lastModified = item.lastModified();
-            if (lastModified > 0)
-            {
-                buf.append(dfmt.format(new Date(item.lastModified())));
-            }
-            buf.append("&nbsp;</td>");
-
-            // Size
-            buf.append("<td class=\"size\">");
-            long length = item.length();
-            if (length >= 0)
-            {
-                buf.append(String.format("%,d bytes", item.length()));
-            }
-            buf.append("&nbsp;</td></tr>\n");
-        }
-        buf.append("</tbody>\n");
-        buf.append("</table>\n");
-        buf.append("</body></html>\n");
-
-        return buf.toString();
-    }
-
-    /**
      * Get the raw (decoded if possible) Filename for this Resource.
      * This is the last segment of the path.
      *
@@ -1008,70 +774,6 @@ public abstract class Resource implements ResourceFactory
         }
 
         return null;
-    }
-
-    /**
-     * Encode any characters that could break the URI string in an HREF.
-     * Such as <a href="/path/to;<script>Window.alert("XSS"+'%20'+"here");</script>">Link</a>
-     *
-     * The above example would parse incorrectly on various browsers as the "<" or '"' characters
-     * would end the href attribute value string prematurely.
-     *
-     * @param raw the raw text to encode.
-     * @return the defanged text.
-     */
-    private static String hrefEncodeURI(String raw)
-    {
-        StringBuffer buf = null;
-
-        loop:
-        for (int i = 0; i < raw.length(); i++)
-        {
-            char c = raw.charAt(i);
-            switch (c)
-            {
-                case '\'':
-                case '"':
-                case '<':
-                case '>':
-                    buf = new StringBuffer(raw.length() << 1);
-                    break loop;
-                default:
-                    break;
-            }
-        }
-        if (buf == null)
-            return raw;
-
-        for (int i = 0; i < raw.length(); i++)
-        {
-            char c = raw.charAt(i);
-            switch (c)
-            {
-                case '"':
-                    buf.append("%22");
-                    break;
-                case '\'':
-                    buf.append("%27");
-                    break;
-                case '<':
-                    buf.append("%3C");
-                    break;
-                case '>':
-                    buf.append("%3E");
-                    break;
-                default:
-                    buf.append(c);
-                    break;
-            }
-        }
-
-        return buf.toString();
-    }
-
-    private static String deTag(String raw)
-    {
-        return StringUtil.sanitizeXmlString(raw);
     }
 
     /**
